+++ app/model/schema/notification.py
@@ -25,7 +25,7 @@ from typing import (
 
 from pydantic import BaseModel
 
-from .result_set import ResultSet
+from .types import ResultSet
 
 
 ############################
